Updates
Scripts now provides the following:

PRs with feature to release, the release branches are updated as well. As it was only updating PRs with release to master.

Should there be a conflict after an update, the bot will inform the author of the PR that there is a conflict and instructions at how to tackle it.

Should a feature to release PR be open for a while, and master is updated multiple times. To avoid duplicate instructions to tackle the conflict, it will mitigate this.